Refactor/session service

+++ src/main/java/br/healthx/Healthx/session/model/service/SessionService.java
@@ -1,9 +1,12 @@
 package br.healthx.Healthx.session.model.service;
 
 import java.time.LocalDate;
-import java.util.List;
+import java.util.NoSuchElementException;
 
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import br.healthx.Healthx.session.dto.SessionRequestDTO;
 import br.healthx.Healthx.session.dto.SessionResponseDTO;
@@ -14,10 +17,10 @@ import br.healthx.Healthx.session.model.entity.Status;
 import br.healthx.Healthx.session.model.exception.ResourceNotFoundException;
 import br.healthx.Healthx.session.model.repository.SessionRepository;
 import br.healthx.Healthx.session.model.validation.SessionValidation;
-import jakarta.transaction.Transactional;
+import lombok.extern.slf4j.Slf4j;
 
+@Slf4j
 @Service
-@Transactional
 public class SessionService {
 
     private final SessionRepository sessionRepository;
@@ -31,56 +34,105 @@ public class SessionService {
         this.sessionValidation = sessionValidation;
     }
 
+    @Transactional(timeout = 30)
     public SessionResponseDTO createSession(SessionRequestDTO dto) {
 
-        sessionValidation.validateSessionCreation(dto);
+        log.info("Attempt to create an session");
+
+        sessionValidation.validateSession(dto);
 
         Session session = sessionMapper.DtoToSession(dto);
 
         sessionRepository.save(session);
 
+        log.info("Session with ID: {} created successfully", session.getId());
+
         return sessionMapper.sessionToDTO(session);
     }
 
+    @Transactional(timeout = 30)
     public SessionResponseDTO updateSession(Long id, SessionRequestDTO dto) {
-        sessionValidation.validateSessionCreation(dto);
+        log.info("Trying to update a session of id: {}", id);
+
+        sessionValidation.validateSession(dto);
         Session session = sessionRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Session not found with id: " + id));
 
-        session = sessionRepository.save(sessionMapper.mapperUpateSession(session, dto));
+        session = sessionRepository.save(sessionMapper.mapperUpdateSession(session, dto));
+
+        log.info("Session of id: {} updated sucessfully", id);
 
         return sessionMapper.sessionToDTO(session);
     }
 
-    public void deleteSession(Long id) {
+    @Transactional(timeout = 30)
+    public void deleteSession(Long id) throws NoSuchElementException {
+        log.info("Trying to delete session of ID: {} ", id);
+
+        if (sessionValidation.ExistSessionId(id)) {
+            throw new ResourceNotFoundException("The Session of id : " + id + " does not exist");
+        }
+
         sessionRepository.deleteById(id);
+
+        log.info("Session of ID: {} deleted sucessfully", id);
+
     }
 
+    @Transactional(readOnly = true, timeout = 15)
     public SessionResponseDTO findSession(Long id) {
+
         Session session = sessionRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Session not found with id: " + id));
+
+        log.info("Trying to fetch a session with the ID: {} ", id);
 
         return sessionMapper.sessionToDTO(session);
     }
 
-    public List<SessionResponseDTO> findAllSessions() {
-        return sessionMapper.sessionToListDTO(sessionRepository.findAll());
+    @Transactional(readOnly = true, timeout = 15)
+    public Page<SessionResponseDTO> findAllSessions(Pageable pageable) {
+
+        log.debug("SessionService:findAllSession fetching all session");
+
+        return sessionMapper.sessionToListDTO(sessionRepository.findAll(pageable));
+
     }
 
-    public List<SessionResponseDTO> findByStatus(Status status) {
-        return sessionMapper.sessionToListDTO(sessionRepository.findByStatus(status));
+    @Transactional(readOnly = true, timeout = 15)
+    public Page<SessionResponseDTO> findByStatus(Status status, Pageable pageable) {
+        if (sessionValidation.validateFields(status))
+            throw new ResourceNotFoundException("This Resource does not exist");
+
+        log.info("Trying to fetch A session by status {} ", status);
+        return sessionMapper.sessionToListDTO(sessionRepository.findByStatus(status, pageable));
     }
 
-    public List<SessionResponseDTO> findBySessionType(SessionType sessionType) {
-        return sessionMapper.sessionToListDTO(sessionRepository.findBySessionType(sessionType));
+    @Transactional(readOnly = true, timeout = 15)
+    public Page<SessionResponseDTO> findBySessionType(SessionType sessionType, Pageable pageable) {
+        if (sessionValidation.validateFields(sessionType))
+            throw new ResourceNotFoundException("This Resource does not exist");
+
+        log.info("Trying to fetch a session containing the sessionType: {}", sessionType);
+        return sessionMapper.sessionToListDTO(sessionRepository.findBySessionType(sessionType, pageable));
     }
 
-    public List<SessionResponseDTO> findSessionByName(String name) {
-        return sessionMapper.sessionToListDTO(sessionRepository.findByPatient_NameContainingIgnoreCase(name));
+    @Transactional(readOnly = true, timeout = 15)
+    public Page<SessionResponseDTO> findSessionByName(String name, Pageable pageable) {
+        if (sessionValidation.validateFields(name))
+            throw new ResourceNotFoundException("This Resource does not exit");
+
+        log.info("Trying to fetch a session that contain the patient name: {} ", name);
+        return sessionMapper.sessionToListDTO(sessionRepository.findByPatient_NameContainingIgnoreCase(name, pageable));
     }
 
-    public List<SessionResponseDTO> findByStartDate(LocalDate start, LocalDate end) {
-        return sessionMapper.sessionToListDTO(sessionRepository.findByStartDateBetween(start, end));
+    @Transactional(readOnly = true, timeout = 15)
+    public Page<SessionResponseDTO> findByStartDate(LocalDate start, LocalDate end, Pageable pageable) {
+        if (sessionValidation.validateFields(start, end))
+            throw new ResourceNotFoundException("This Resource does not exit");
+
+        log.info("Trying to fetch Session that is between the startDate: {} and the endDate {}", start, end);
+        return sessionMapper.sessionToListDTO(sessionRepository.findByStartDateBetween(start, end, pageable));
     }
 
 }
